backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "escrow-bridge"
version = "0.1.0"
description = "Python Library for Escrow Bridge"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"

authors = [
    { name = "Brandyn Hamilton", email = "brandynham1120@gmail.com"}
]
dependencies = [
    "click>=8.1.8",
    "diskcache>=5.6.3",
    "fastapi>=0.103.2",
    "httpx>=0.24.1",
    "jinja2>=3.1.6",
    "python-dotenv>=0.21.1",
    "uvicorn>=0.22.0",
    "web3>=5.3.1",
    "chartengineer==0.1.5",
    "apscheduler>=3.10.4",
    "rich>=13.8.1",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "bcrypt>=4.2.1",
]

[tool.setuptools.packages.find]
include = ["escrow_bridge", "escrow_bridge.*"]

[project.scripts]
escrow-bridge = "escrow_bridge.cli.main:cli"
escrow-bridge-admin = "escrow_bridge.cli.admin:cli"
